+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Next
+
+* Fixed missing .gitignore in project template
+
 ## 1.0.2 (2024-03-10)
 
 * Project generator available.